chapter-03/requirements.txt
wheel
Flask-Appbuilder
Flask-Migrate
mysqlclient
google-cloud-texttospeech
